Move EVERYONE_ORG_ID into common/consts.ts
We used to have it in client and server consts separately, move it
into the common file.
