+++ src/common/consts.ts
@@ -1,1 +1,2 @@
 export const DM_CHANNEL_PREFIX = 'dm:';
+export const EVERYONE_ORG_ID = 'clack_all';
