+++ src/server/handlers/getCordThreads.ts
@@ -1,7 +1,7 @@
 import type { CoreThreadData } from '@cord-sdk/types';
 import type { Request, Response } from 'express';
 import { fetchCordRESTApi } from 'src/server/fetchCordRESTApi';
-import { EVERYONE_ORG_ID } from 'src/server/consts';
+import { EVERYONE_ORG_ID } from 'src/common/consts';
 
 export async function handleGetMyCordThreads(_: Request, res: Response) {
   const allThreads = await fetchCordRESTApi<CoreThreadData[]>('threads');
